fix(doom): fold in code-review hardening
- cv_to_norm guards a zero/negative denominator (deadzone >= full scale) instead
  of relying on IEEE inf-comparison to saturate.
- seg_sidedef_index and seg_is_one_sided bounds-check segIndex itself, matching
  the other hardened resolvers.
- real_wad_probe uses std::cos/std::sin (cosf/sinf are not guaranteed in std).

+++ harness/tools/real_wad_probe.cpp
@@ -24,7 +24,7 @@
 #include "../../plugins/games/doom/fb.h"
 #include "wav_wrap.h"
 
-namespace doom { void cos_sin(float a, float& c, float& s) { c = std::cosf(a); s = std::sinf(a); } }
+namespace doom { void cos_sin(float a, float& c, float& s) { c = std::cos(a); s = std::sin(a); } }
 
 static std::vector<uint8_t> read_file(const char* path) {
     FILE* f = fopen(path, "rb");
